+++ packages/plasma-b2c/src/components/AudioPlayer/index.ts
@@ -1,0 +1,2 @@
+export { AudioPlayer } from './AudioPlayer';
+export type { AudioPlayerProps } from './AudioPlayer';