+++ packages/plasma-b2c/src/components/PreviewGallery/index.ts
@@ -1,0 +1,2 @@
+export { PreviewGallery } from './PreviewGallery';
+export { arrayItemRemoving, arrayItemSelecting, arrayItemSwapping } from './utils';