+++ packages/plasma-ui/CHANGELOG.md
@@ -1,3 +1,30 @@
+# v1.57.0 (Thu Oct 21 2021)
+
+#### 🚀 Enhancement
+
+- feat(plasma-ui): `Header`'s back-minimize animation" [#863](https://github.com/sberdevices/plasma/pull/863) ([@fanisco](https://github.com/fanisco))
+- feat(plasma-ui): Added `NueHeader` to replace `Header` in major version [#863](https://github.com/sberdevices/plasma/pull/863) ([@fanisco](https://github.com/fanisco))
+
+#### Authors: 1
+
+- Fanil' Zubairov ([@fanisco](https://github.com/fanisco))
+
+---
+
+# v1.56.1 (Thu Oct 21 2021)
+
+#### 🐛 Bug Fix
+
+- fix(plasma-ui): `HeaderContent` flex in Chrome 94 [#875](https://github.com/sberdevices/plasma/pull/875) ([@fanisco](https://github.com/fanisco))
+- chore: update package-locks \[skip ci\] ([@SberDevices-Eva](https://github.com/SberDevices-Eva))
+
+#### Authors: 2
+
+- [@SberDevices-Eva](https://github.com/SberDevices-Eva)
+- Fanil' Zubairov ([@fanisco](https://github.com/fanisco))
+
+---
+
 # v1.55.3 (Fri Oct 15 2021)
 
 #### 🐛 Bug Fix